refactor(attributes): introduce Preact shell for attribute detail
Rename the legacy widget to attribute_detail_old.ts and replace it with a
transitional adapter keeping the showAttributeDetail()/hide() contract while
the UI is ported piece by piece to a Preact component.
